MultiVal attributes call __setattr__ of instance
MultiValAttribute and QualifiedMultiValAttribute now call Atrribut.__set__
(and this way __setattr__ of containing instance) when their container is
modified. This also happens, when a default value is returned and then
modified, as long as the defined default value is is not callable. This is
achieved by associating the returned default value with the instance when
accessed via the attributes __get__ method.