Changed adapter lookup.
If the lookup does not find an adapter among the adapter registered with the
asked component, it now proceeds with searching the subclasses of that
component. If more than one adapter is registered for the type of the object
to be adapted, the adapter which was registered the latest is returned.